refactor: ExhibitionSearchBar 컴포넌트를 모듈화
